Add history query API and persistent implementation (#151)
This refactors the history module, adds fuller interface specification for the get call and propagates that to the web app. Also implements a more useful persistent history backed by partitioned json files on disk, that isolates data in "experiments", in addition to the demo in memory history implementation that can be used in testing.