experiment layer (#244)
Place a grouping layer between the device manager and the experiments, which are represented as a set of controllers. This layer provides maximal flexibility in efficiently utilizing hardware and maps to the way evolver is typically used. 

An experiment route is also added along with modifications to history to support a useful experiments endpoint grouping data from all controllers.