Store fitted calibrator in state model, path to load, and default transformers (#268)
Add facility to store fitted calibrator in calibration data file and load on initialization. Also adding default transformers to the per vial calibrator config for cases where all the same class and config (outside fitted parameters) is used and refit is expected either in the procedure or in the initialization.