making all things deserialize as configdesc, iso config (#295)
This makes BaseInterface deserialize to ConfigDescriptors, enabling intended types usable in the model annotations and work ask expected. This also makes them symmetric with how the serialize (as ConfigDescriptors). Union annotations with ConfigDescriptor are then redundant and so are removed in this commit.

Additionally this upgrades pydantic to the latest version in tests and removes the version limit.